+++ src/components/Radio/index.js
@@ -24,6 +24,7 @@ const Radio = (props) => {
           name={ props.name }
           defaultChecked={ props.defaultChecked }
           checked={ props.checked }
+          className="flex--none"
           disabled={ props.isDisabled }
           onChange={ props.onChange }
           data-test-section={ props.testSection }